add some casts
git-svn-id: https://svn.r-project.org/R/trunk@59115 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/stats/src/massdist.c
@@ -51,7 +51,7 @@ void massdist(double *x, double *xmass, /* AB: new variable */
         if (R_FINITE(x[i]))
         {
             xpos = (x[i] - *xlow) / xdelta;
-            ix = floor(xpos);
+            ix = (int)floor(xpos);
             fx = xpos - ix;
             xmi = xmass[i]; /* AB: new line  */
             if (ixmin <= ix && ix <= ixmax)
